Preserver mediaqueries
This searches all style tags for Media Queries, and replaces the content with just those. Fixes #45

Also set the default to strip mediaqueries from inline tags, because that doesn't work as expected. (at least, not after #52)

Not sure if this is the most performant/optimal way, so open for ideas ;)