Refactor removing style tags
This uses xPath to find the style tags and either remove the style tag completely, or replace them with just the media queries, based on the excludeMediaQueries flag.
This follows the logic that if the mediaqueries are excluded from inlining, they should not be stripped from the document, because they aren't processed.
If for some reason they are processed (don't know the use case) they can just be removed.