Reuse the same CssSelectorConverter
Initializing the CSS converter is a bit heavy. Reusing the same instance
for all rules instead of creating a new one each time leads to a major
performance boost (-20% time spent when running the example).
This optimization is possible only when using the Symfony CssSelector
component 2.8+. The old API from Symfony 2.7 and older does not allow to
perform it, as it exposes only a static method.