Correct order for inline styles
Place styles that were inline in the source styles after the styles that are being inlined. This is because some properties may be set using different names. Example:

Example:

    <!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
    <html xmlns="http://www.w3.org/1999/xhtml" lang="en" xml:lang="en">
       <head>
          <meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
          <meta name="viewport" content="width=device-width"/>
          <title>Test css inliner</title>
          <style type="text/css">
    th {
       background: red;
    }
          </style>
       </head>
       <body>
          <table>
             <tr>
                <th style="background-color: blue;">This should be blue</th>
             </tr>
          </table>
       </body>
    </html>

Will result in `<th style="background-color: blue; background: red;">This should be blue</th>` when the inlined styles are positioned after the styles that were already inline.

This issue is fixed by inlining styles before the styles that were already inline.