Remove unnecessary cleanup method in tests
Removing the reference to the CssToInlineStyles instance after each test
is unnecessary as the TestCase itself will be collected once not
referenced anymore (and there is no circular reference preventing the
refcount to collect them).
Removing this method avoids getting static analysis reports about
assigning `null` in a non-nullable property.